# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Scrapor API
  version: 1.0.0
  description: |
    Scrapor multi-tenant video platform API.

    ## Authentication

    All documented endpoints accept an API key via:
    - Header: `Authorization: Bearer spk_ro_{id}_{secret}` or `spk_fa_{id}_{secret}`
    - Query: `?apiKey=spk_…`

    **Key types:**
    - `READ_ONLY` (`spk_ro_…`) — GET/HEAD/OPTIONS only, including full video stream via `GET /api/videos/{id}/stream`
    - `FULL_ACCESS` (`spk_fa_…`) — all HTTP methods on all routes (except `/api/api-keys/*` which is admin-session only)

    Browser sessions (cookies) also work for the admin panel and public site. When an API key is present, its type and method permissions are enforced.

    This spec and `/api/docs` require a valid API key (any type).
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token with format `spk_ro_{keyNumber}_{secret}` or `spk_fa_{keyNumber}_{secret}`.
        Alternatively pass `apiKey` as a query parameter.
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/videos/{id}/vast:
    post:
      tags: [videos]
      summary: Resolve VAST ad tag
      description: Server-side VAST resolution for the video player.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                adSessionId: { type: string }
      responses:
        "200": { description: VAST ad payload or null }
...
  /api/creator/videos:
    post:
      tags: [creator]
      summary: Upload creator video
      description: Streams raw video body to disk. Metadata via query params. Requires creator session or FULL_ACCESS key with session.
      parameters:
        - name: title
          in: query
          required: true
          schema: { type: string }
        - name: description
          in: query
          schema: { type: string }
        - name: tags
          in: query
          schema: { type: string }
        - name: durationSec
          in: query
          schema: { type: integer }
      responses:
        "200": { description: Upload accepted }
        "403": { description: Creator profile required }
...